more test fixing
